+++ package/Framework/Interactive Tutorials/TutorialInfo/Scripts/Editor/ReadmeEditor.cs
@@ -54,11 +54,8 @@ public class ReadmeEditor : Editor {
 			
 			return (Readme)readmeObject;
 		}
-		else
-		{
-			Debug.Log("Couldn't find a readme");
-			return null;
-		}
+
+		return null;
 	}
 	
 	protected override void OnHeaderGUI()